% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{create_ec_model}
\alias{create_ec_model}
\title{Create EC Model}
\usage{
create_ec_model(ec_annotation)
}
\arguments{
\item{ec_annotation}{A dataframe with EC annotation information}
}
\value{
A list of dataframes with EC pathway background frequencies per EC pathway type
}
\description{
This function generates a background frequency model for EC pathway enrichment analysis.
}
\examples{
ec_model_df <- create_ec_model(ec_annotation_df)
}
