% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{create_go_model}
\alias{create_go_model}
\title{Create GO Model}
\usage{
create_go_model(go_annotation)
}
\arguments{
\item{go_annotation}{A dataframe with GO annotation information}
}
\value{
A list of dataframes with GO terms background frequencies per GO term type
}
\description{
This function generates a background frequency model for GO term enrichment analysis.
}
\examples{
go_model_df <- create_go_model(go_annotation_df)
}
