% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{create_go_model_eggnog}
\alias{create_go_model_eggnog}
\title{Create GO Model from eggNOG-mapper Annotation}
\usage{
create_go_model_eggnog(eggnog_annotation)
}
\arguments{
\item{eggnog_annotation}{A dataframe with eggNOG-mapper annotation information (see load_eggnog_annotation)}
}
\value{
A list of dataframes with GO terms background frequencies by gotermType (biological_processes, cellular_component, molecular_function)
}
\description{
This function generates a background frequency model for GO term enrichment analysis based on eggNOG-mapper annotation.
}
\examples{
go_model_df <- create_go_model_eggnog(eggnog_annotation_df)
go_model_biological_processes_df <- go_model_df$biological_processes
go_model_cellular_component_df <- go_model_df$cellular_component
go_model_molecular_function_df <- go_model_df$molecular_function
}
