% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v5.R
\name{create_kegg_model}
\alias{create_kegg_model}
\title{Create KEGG Pathway Model}
\usage{
create_kegg_model(kegg_annotation, filter = "All")
}
\arguments{
\item{kegg_annotation}{A dataframe with KEGG annotation information}

\item{filter}{A filter for KEGG pathway annotation (Kingdom, Phylum, Class, Organism)(default = "All")}
}
\value{
A list of dataframes with KEGG pathway background frequencies per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete)
}
\description{
This function generates a background frequency model for KEGG pathway enrichment analysis.
}
\examples{
kegg_model_df <- create_kegg_model(kegg_annotation_df)
}
