% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v5.R
\name{create_kegg_model_eggnog}
\alias{create_kegg_model_eggnog}
\title{Create KEGG Pathway Model from eggNOG-mapper Annotation}
\usage{
create_kegg_model_eggnog(eggnog_annotation, filter = "All")
}
\arguments{
\item{eggnog_annotation}{A dataframe with eggNOG-mapper annotation information}

\item{filter}{A filter for KEGG pathway annotation (Kingdom, Phylum, Class, Organism)(default = "All")}
}
\value{
A list of dataframes with KEGG pathway background frequencies per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete).
}
\description{
This function generates a background frequency model for KEGG pathway enrichment analysis based on eggNOG-mapper annotation.
}
\examples{
kegg_model_df <- create_kegg_model_eggnog(eggnog_annotation_df)
}
