% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{create_kog_model}
\alias{create_kog_model}
\title{Create COG/KOG Model}
\usage{
create_kog_model(kog_annotation)
}
\arguments{
\item{kog_annotation}{A dataframe with COG/KOG annotation information (see load_kog_annotation)}
}
\value{
A dataframe with COG/KOG categories background frequencies
}
\description{
This function generates a background frequency model for COG/KOG categories enrichment analysis from a COG/KOG annotation.
}
\examples{
kog_model_df <- create_kog_model(kog_annotation_df)
}
