% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{create_kog_model_eggnog}
\alias{create_kog_model_eggnog}
\title{Create COG/KOG Model from eggNOG-mapper Annotation}
\usage{
create_kog_model_eggnog(eggnog_annotation)
}
\arguments{
\item{eggnog_annotation}{A dataframe with eggNOG-mapper annotation information (see load_eggnog_annotation)}
}
\value{
A dataframe with COG/KOG categories background frequencies
}
\description{
This function generates a background frequency model for COG/KOG categories enrichment analysis based on eggNOG-mapper annotation.
}
\examples{
kog_model_df <- create_kog_model_eggnog(eggnog_annotation_df)
}
