% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{create_transcript2protein_id_df}
\alias{create_transcript2protein_id_df}
\title{Create Transcript ID to Protein ID Dataframe}
\usage{
create_transcript2protein_id_df(gtf_df)
}
\arguments{
\item{gtf_df}{A dataframe with GTF annotation information}
}
\value{
A dataframe with transcript ID and protein ID information
}
\description{
This function parses a GTF annotation dataframe to generate a transcript ID to protein ID conversion dataframe.
}
\examples{
transcript2protein_id_df <- create_transcript2protein_id_df(gtf_df)
}
