% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{ec_enrichment}
\alias{ec_enrichment}
\title{Create EC Enrichment Dataframe}
\usage{
ec_enrichment(
  ec_model,
  ec_annotation_df,
  protein_ids,
  test = "fisher",
  p.adjust.method = "BH"
)
}
\arguments{
\item{ec_model}{A list of dataframes with EC pathway background frequencies per EC pathway type}

\item{ec_annotation_df}{A dataframe with EC annotation information}

\item{protein_ids}{A vector of protein IDs}

\item{test}{A statistical method for testing the null of independence of rows and columns in a contingency table (default = "fisher")}

\item{p.adjust.method}{A statistical method to adjust p-values for multiple comparisons (default = "BH")}
}
\value{
A list of dataframes with EC pathway enrichment statistics per EC pathway type
}
\description{
This function generates a dataframe containing background/sample frequencies and enrichment statistics per EC pathway type.
}
\examples{
ec_enrichment_df <- ec_enrichment(ec_model_df, ec_annotation_df, protein_ids)
ec_enrichment_df <- ec_enrichment(ec_model_df, ec_annotation_df, protein_ids, test= "fisher", p.adjust.method= "BH")
ec_enrichment_df <- ec_enrichment(ec_model_df, ec_annotation_df, protein_ids, test= "chisq", p.adjust.method= "bonferonni")
ec_enrichment_df <- ec_enrichment(ec_model_df, ec_annotation_df, protein_ids, test= "chisq", p.adjust.method= "none")
}
