% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{fetch_models}
\alias{fetch_models}
\title{Fetch Models}
\usage{
fetch_models(strain)
}
\arguments{
\item{enrichment_df}{A list of dataframes with KEGG pathway enrichment statistics per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete)}
}
\value{
A list with model information, gtf annotation and precomputed KOG/GO/KEGG models (info, gtf, transcript2protein_id, kog, go, kegg)
}
\description{
This function loads precomputed KOG/GO/KEGG models for an organism and print infos on data origin
}
\examples{
qm6a_models <- fetch_models("Trichoderma reesei (QM6a)")
qm6a_gtf <- qm6a_models$gtf
qm6a_transcript2protein_id <- qm6a_models$transcript2protein_id
qm6a_kog <- qm6a_models$kog
qm6a_go <- qm6a_models$go
qm6a_kegg <- qm6a_models$kegg
}
