% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{gene2protein_id}
\alias{gene2protein_id}
\title{Gene ID to Protein ID}
\usage{
gene2protein_id(genes, transcript2protein_id_df)
}
\arguments{
\item{transcript2protein_id_df}{A dataframe with gene ID, transcript ID and protein ID information}

\item{transcript_ids}{A vector of gene IDs}
}
\value{
A vector of protein IDs
}
\description{
Converts a list of Gene IDs and/or Gene Names to Protein Ids.
}
\examples{
protein_ids <- gene2protein_id(genes, transcript2protein_id_df)
}
