% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{generate_ec_plot}
\alias{generate_ec_plot}
\title{Generate EC Pathway Enrichment Plot}
\usage{
generate_ec_plot(
  enrichment_df,
  pathwayType = "summary",
  n = 10,
  significant = "#9986A5",
  plot_title = NA
)
}
\arguments{
\item{enrichment_df}{A list of dataframes with EC pathway enrichment statistics per EC pathway type}

\item{pathwayType}{A EC pathway type ("summary", "regulatory" or "metabolic"; default = "summary")}

\item{n}{A number of top-ranking EC pathways to include in the plot (default = 10)}

\item{significant}{A color assignment for statistically significant GO terms (default = "#9986A5")}

\item{plot_title}{A plot title (default = NA)}
}
\value{
A ggplot2 bar plot object
}
\description{
This function generates a bar plot for visualization of EC pathways enrichment statistics.
}
\examples{
generate_ec_plot(ec_enrichment_df)
generate_ec_plot(ec_enrichment_df, pathwayType = "summary", plot_title = "EC Enrichment: Summary")
generate_ec_plot(ec_enrichment_df, pathwayType = "metabolic", significant = "blue")
generate_ec_plot(ec_enrichment_df, pathwayType = "regulatory", n = 15)
}
