% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{generate_kegg_plot}
\alias{generate_kegg_plot}
\title{Generate KEGG Pathway Enrichment Plot}
\usage{
generate_kegg_plot(
  enrichment_df,
  model = "pathway_class",
  n = 10,
  significant = "#9986A5",
  plot_title = NA,
  plot_type = "bar"
)
}
\arguments{
\item{enrichment_df}{A list of dataframes with KEGG pathway enrichment statistics per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete) (see kegg_enrichment)}

\item{n}{A number of top-ranking KEGG pathways to include in the plot (default = 10)}

\item{significant}{A color assignment for statistically significant GO terms (default = "#9986A5")}

\item{plot_title}{A plot title (default = NA)}

\item{plot_type}{A plot type option (default = "bar"; options = c("bar", "lollipop"))}

\item{pathwayType}{A KEGG category for plotting (default = "pathway_class"; options = c("definition", "pathway_name", "pathway_class", "pathway_type"))}
}
\value{
A ggplot2 plot object
}
\description{
This function generates a bar plot for visualization of KEGG pathway enrichment statistics.
}
\examples{
generate_kegg_plot(kegg_enrichment_df)
generate_kegg_plot(kegg_enrichment_df, model = "pathway_class", plot_title = "KEGG Pathway Enrichment: Class")
generate_kegg_plot(kegg_enrichment_df, model = "pathway_name", significant = "blue")
generate_kegg_plot(kegg_enrichment_df, model = "pathway_name", n = 5, plot_type = "lollipop")
}
