% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{generate_kog_plot}
\alias{generate_kog_plot}
\title{Generate COG/KOG Enrichment Plot}
\usage{
generate_kog_plot(
  enrichment_df,
  significant = "#0B775E",
  plot_title = NA,
  plot_type = "bar"
)
}
\arguments{
\item{enrichment_df}{A dataframe with COG/KOG enrichment statistics (see kog_enrichment)}

\item{significant}{A color assignment for statistically significant COG/KOG categories (default = "#0B775E")}

\item{plot_title}{A plot title (default = NA)}

\item{plot_type}{A plot type option (default = "bar"; options = c("bar", "lollipop"))}
}
\value{
A ggplot2 bar plot object
}
\description{
This function generates a bar plot for visualization of COG/KOG categories enrichment statistics.
}
\examples{
generate_kog_plot(kog_enrichment_df)
generate_kog_plot(kog_enrichment_df, plot_type = "lollipop")
generate_kog_plot(kog_enrichment_df, plot_title = "KOG Enrichment", significant = "blue")
}
