% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{go_enrichment}
\alias{go_enrichment}
\title{Calculate GO Enrichment}
\usage{
go_enrichment(go_model, protein_ids, test = "fisher", p.adjust.method = "BH")
}
\arguments{
\item{go_model}{A list of dataframes with GO terms background frequencies per GO term type}

\item{protein_ids}{A vector of protein IDs}

\item{test}{A statistical method for testing the null of independence of rows and columns in a contingency table (default = "fisher")}

\item{p.adjust.method}{A statistical method to adjust p-values for multiple comparisons (default = "BH")}
}
\value{
A list of dataframes with GO terms enrichment statistics per GO term type
}
\description{
This function generates a dataframe containing background/sample frequencies and enrichment statistics per GO term.
}
\examples{
go_enrichment_df <- go_enrichment(go_model_df, protein_ids)
go_enrichment_df <- go_enrichment(go_model_df, protein_ids, test = "fisher", p.adjust.method = "BH")
go_enrichment_df <- go_enrichment(go_model_df, protein_ids, test = "chisq", p.adjust.method = "bonferroni")
go_enrichment_df <- go_enrichment(go_model_df, protein_ids, test = "chisq", p.adjust.method = "none")
}
