% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{go_enrichment_eggnog}
\alias{go_enrichment_eggnog}
\title{Create GO Enrichment Dataframe from eggNOG-mapper Annotation}
\usage{
go_enrichment_eggnog(
  go_model,
  eggnog_annotation,
  protein_ids,
  test = "fisher",
  p.adjust.method = "BH"
)
}
\arguments{
\item{go_model}{A list of dataframes with GO terms background frequencies per GO term type}

\item{eggnog_annotation}{A dataframe with eggNOG-mapper annotation information}

\item{protein_ids}{A vector of protein IDs}

\item{test}{A statistical method for testing the null of independence of rows and columns in a contingency table (default = "fisher")}

\item{p.adjust.method}{A statistical method to adjust p-values for multiple comparisons (default = "BH")}
}
\value{
A list of dataframes with GO terms enrichment statistics per GO term type (biological_processes, cellular_component, molecular_function)
}
\description{
This function generates a dataframe containing background/sample frequencies and enrichment statistics per GO term from eggNOG-mapper Annotation.
}
\examples{
go_enrichment_df <- go_enrichment(go_model_df, eggnog_annotation, protein_ids)
go_enrichment_df <- go_enrichment(go_model_df, eggnog_annotation, protein_ids, test = "fisher", p.adjust.method = "BH")
go_enrichment_df <- go_enrichment(go_model_df, eggnog_annotation, protein_ids, test = "chisq", p.adjust.method = "bonferroni")
go_enrichment_df <- go_enrichment(go_model_df, eggnog_annotation, protein_ids, test = "chisq", p.adjust.method = "none")
go_enrichment_biological_processes_df <- go_enrichment_df$biological_processes
go_enrichment_cellular_component_df <- go_enrichment_df$cellular_component
go_enrichment_molecular_function_df <- go_enrichment_df$molecular_function
}
