% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{kegg_enrichment}
\alias{kegg_enrichment}
\title{Calculate KEGG Pathway Enrichment}
\usage{
kegg_enrichment(
  kegg_model,
  protein_ids,
  test = "fisher",
  p.adjust.method = "BH"
)
}
\arguments{
\item{kegg_model}{A list of dataframes with KEGG pathway background frequencies per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete)}

\item{protein_ids}{A vector of protein IDs}

\item{test}{A statistical method for testing the null of independence of rows and columns in a contingency table (default = "fisher")}

\item{p.adjust.method}{A statistical method to adjust p-values for multiple comparisons (default = "BH")}
}
\value{
A list of dataframes with KEGG pathways enrichment statistics per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete)
}
\description{
This function generates a list of dataframes containing background/sample frequencies and enrichment statistics per KEGG category.
}
\examples{
kegg_enrichment_df <- kegg_enrichment(kegg_model_df, protein_ids)
kegg_enrichment_df <- kegg_enrichment(kegg_model_df, protein_ids, test = "fisher", p.adjust.method = "BH")
kegg_enrichment_df <- kegg_enrichment(kegg_model_df, protein_ids, test = "chisq", p.adjust.method = "bonferroni")
kegg_enrichment_df <- kegg_enrichment(kegg_model_df, protein_ids, test = "chisq", p.adjust.method = "none")
}
