% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{kog_enrichment}
\alias{kog_enrichment}
\title{Calculate COG/KOG Enrichment}
\usage{
kog_enrichment(kog_model, protein_ids, test = "fisher", p.adjust.method = "BH")
}
\arguments{
\item{kog_model}{A dataframe with COG/KOG categories background frequencies}

\item{protein_ids}{A vector of protein IDs}

\item{test}{A statistical method for testing the null of independence of rows and columns in a contingency table (default = "fisher")}

\item{p.adjust.method}{A statistical method to adjust p-values for multiple comparisons (default = "BH")}
}
\value{
A dataframe with COG/KOG categories enrichment statistics
}
\description{
This function generates a dataframe containing background/sample frequencies and enrichment statistics per COG/KOG category.
}
\examples{
kog_enrichment_df <- kog_enrichment(kog_model_df, protein_ids)
kog_enrichment_df <- kog_enrichment(kog_model_df, protein_ids, test = "fisher", p.adjust.method = "BH")
kog_enrichment_df <- kog_enrichment(kog_model_df, protein_ids, test = "chisq", p.adjust.method = "bonferroni")
kog_enrichment_df <- kog_enrichment(kog_model_df, protein_ids, test = "chisq", p.adjust.method = "none")
}
