% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{load_custom_ec_annotation}
\alias{load_custom_ec_annotation}
\title{Load Custom EC Annotation}
\usage{
load_custom_ec_annotation(path)
}
\arguments{
\item{path}{Path to EC annotation}
}
\value{
A dataframe with EC annotation information
}
\description{
This function loads a custom EC (Enzyme Commission) annotation into a dataframe.
}
\examples{
ec_annotation_df <- load_custom_ec_annotation(ec_annotation_path)
}
