% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{load_custom_go_annotation}
\alias{load_custom_go_annotation}
\title{Load Custom GO Annotation}
\usage{
load_custom_go_annotation(path)
}
\arguments{
\item{path}{Path to GO annotation}
}
\value{
A dataframe with GO annotation information
}
\description{
This function loads a custom GO (Gene Ontology) annotation into a dataframe.
}
\examples{
go_annotation_df <- load_custom_go_annotation(go_annotation_path)
}
