% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{load_custom_kog_annotation}
\alias{load_custom_kog_annotation}
\title{Load Custom COG/KOG Annotation}
\usage{
load_custom_kog_annotation(path)
}
\arguments{
\item{path}{Path to COG/KOG annotation}
}
\value{
A dataframe with COG/KOG annotation information
}
\description{
This function loads a custom COG/KOG (Clusters of Orthologous Genes) annotation into a dataframe.
}
\examples{
kog_annotation_df <- load_custom_kog_annotation(kog_annotation_path)
}
