% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{load_eggnog_annotation}
\alias{load_eggnog_annotation}
\title{Load EggNOG-Mapper Annotation}
\usage{
load_eggnog_annotation(path)
}
\arguments{
\item{path}{Path to eggNOG-mapper annotation file}
}
\value{
A dataframe with eggNOG-mapper annotation information
}
\description{
This functions loads a eggNOG-mapper annotation into a dataframe.
}
\examples{
eggnog_annotation_df <- load_eggnog_annotation(path/to/eggnog-mapper/annotation)
}
