% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v5.R
\name{load_go_annotation}
\alias{load_go_annotation}
\title{Load GO Annotation}
\usage{
load_go_annotation(path)
}
\arguments{
\item{path}{Path to GO annotation}
}
\value{
A dataframe with GO annotation information
}
\description{
This function loads a GO (Gene Ontology) annotation into a dataframe.
}
\examples{
go_annotation_df <- load_go_annotation(go_annotation_path)
}
