% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v5.R
\name{load_kegg_annotation}
\alias{load_kegg_annotation}
\title{Load KEGG Annotation}
\usage{
load_kegg_annotation(path, transcript2protein_id_df)
}
\arguments{
\item{path}{Path to KEGG annotation}

\item{transcript2protein_id_df}{A dataframe with gene ID, transcript ID and protein ID information}
}
\value{
A dataframe with KEGG annotation information
}
\description{
This function loads a KEGG annotation into a dataframe.
}
\examples{
kegg_annotation_df <- load_kegg_annotation(kegg_annotation_path)
}
