% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea_v3.R
\name{transcript2protein_id}
\alias{transcript2protein_id}
\title{Transcript ID to Protein ID}
\usage{
transcript2protein_id(transcript_ids, transcript2protein_id_df)
}
\arguments{
\item{transcript_ids}{A vector of transcript IDs}

\item{transcript2protein_id_df}{A dataframe with transcript ID and protein ID information}
}
\value{
A vector of protein IDs
}
\description{
Converts a list of Transcript IDs to Protein Ids.
}
\examples{
protein_ids <- transcript2protein_id(transcript_ids, transcript2protein_id_df)
}
