% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{create_kegg_model}
\alias{create_kegg_model}
\title{Create KEGG Pathway Model}
\usage{
create_kegg_model(kegg_annotation)
}
\arguments{
\item{kegg_annotation}{A dataframe with KEGG annotation information (see load_kegg_annotation)}
}
\value{
A list of dataframes with KEGG pathway background frequencies per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete)
}
\description{
This function generates a background frequency model for KEGG pathway enrichment analysis.
}
\examples{
kegg_model_df <- create_kegg_model(kegg_annotation_df)
kegg_model_type_df <- kegg_model_df$pathway_type
kegg_model_class_df <- kegg_model_df$pathway_class
kegg_model_name_df <- kegg_model_df$pathway_name
kegg_model_definition_df <- kegg_model_df$definition
kegg_model_complete_df <- kegg_model_df$complete
}
