% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{create_kegg_model_eggnog}
\alias{create_kegg_model_eggnog}
\title{Create KEGG Pathway Model from eggNOG-mapper Annotation}
\usage{
create_kegg_model_eggnog(eggnog_annotation)
}
\arguments{
\item{eggnog_annotation}{A dataframe with eggNOG-mapper annotation information (see load_eggnog_annotation)}
}
\value{
A list of dataframes with KEGG pathway background frequencies per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete).
}
\description{
This function generates a background frequency model for KEGG pathway enrichment analysis based on eggNOG-mapper annotation.
}
\examples{
kegg_model_df <- create_kegg_model_eggnog(eggnog_annotation_df)
}
