% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{create_kegg_model_from_ec}
\alias{create_kegg_model_from_ec}
\title{Create KEGG Pathway Model from EC Annotation}
\usage{
create_kegg_model_from_ec(ec_annotation_df)
}
\arguments{
\item{ec_annotation}{A dataframe with EC annotation information (see load_ec_annotation)}
}
\value{
A list of dataframes with KEGG pathway background frequencies per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete)
}
\description{
This function generates a background frequency model for KEGG pathway enrichment analysis from an EC (Enzyme Commission) Annotation.
}
\examples{
kegg_model_df <- create_kegg_model_from_ec(ec_annotation_df, filter = "All")
}
