% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{gene2protein_id}
\alias{gene2protein_id}
\title{Convert a Vector of Gene IDs to Protein IDs}
\usage{
gene2protein_id(genes, transcript2protein_id_df)
}
\arguments{
\item{transcript2protein_id_df}{A dataframe with gene ID, transcript ID and protein ID information}

\item{transcript_ids}{A vector of gene IDs and/or gene names}
}
\value{
A vector of protein IDs
}
\description{
This function converts a vector of gene IDs and/or gene names to protein IDs.
}
\examples{
protein_ids <- gene2protein_id(genes, transcript2protein_id_df)
}
