% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{gene2protein_id}
\alias{gene2protein_id}
\title{Convert a Vector of Gene IDs to Protein IDs}
\usage{
gene2protein_id(gene_ids, transcript2protein_id_df)
}
\arguments{
\item{gene_ids}{A vector of gene IDs and/or gene names}

\item{transcript2protein_id_df}{A dataframe with gene ID, gene name, transcript ID and protein ID information (see create_transcript2protein_id_df)}
}
\value{
A vector of protein IDs
}
\description{
This function converts a vector of gene IDs and/or gene names to protein IDs, using a transcript2protein_id dataframe constructed from GTF/GFF/GFF3 annotation.
}
\examples{
protein_ids <- gene2protein_id(gene_ids, transcript2protein_id_df)
}
