% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{kegg_enrichment}
\alias{kegg_enrichment}
\title{Calculate KEGG Pathway Enrichment}
\usage{
kegg_enrichment(
  kegg_model,
  protein_ids,
  test = "fisher",
  p.adjust.method = "BH"
)
}
\arguments{
\item{kegg_model}{A list of dataframes with KEGG pathway background frequencies per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete) (see create_kegg_model)}

\item{protein_ids}{A vector of protein IDs. Must match the Protein IDs from the annotation used to build the model.}

\item{test}{A statistical method for testing the null of independence of rows and columns in a contingency table (default = "fisher"; options = c("fisher", "chisq))}

\item{p.adjust.method}{A statistical method to adjust p-values for multiple comparisons (default = "BH"; options = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")(see p.adjust.methods))}
}
\value{
A list of dataframes with KEGG pathways enrichment statistics per KEGG category (pathway_type, pathway_class, pathway_name, definition, complete)
}
\description{
This function generates a list of dataframes containing background/sample frequencies and enrichment statistics per KEGG category.
}
\examples{
kegg_enrichment_df <- kegg_enrichment(kegg_model_df, protein_ids)
kegg_enrichment_df <- kegg_enrichment(kegg_model_df, protein_ids, test = "fisher", p.adjust.method = "BH")
kegg_enrichment_df <- kegg_enrichment(kegg_model_df, protein_ids, test = "chisq", p.adjust.method = "bonferroni")
kegg_enrichment_type_df <- kegg_enrichment_df$pathway_type
kegg_enrichment_class_df <- kegg_enrichment_df$pathway_class
kegg_enrichment_name_df <- kegg_enrichment_df$pathway_name
kegg_enrichment_definition_df <- kegg_enrichment_df$definition
kegg_enrichment_complete_df <- kegg_enrichment_df$complete
}
