% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{load_ec_annotation}
\alias{load_ec_annotation}
\title{Load EC Annotation}
\usage{
load_ec_annotation(path)
}
\arguments{
\item{path}{Path to EC annotation}
}
\value{
A dataframe with EC annotation information
}
\description{
This function loads a EC (Enzyme Commission) based KEGG annotation into a dataframe.
}
\examples{
ec_annotation_df <- load_ec_annotation(path/to/ec/annotation)
}
