% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{load_gtf_annotation}
\alias{load_gtf_annotation}
\title{Load GTF/GFF Annotation}
\usage{
load_gtf_annotation(path)
}
\arguments{
\item{path}{Path to GTF annotation}
}
\value{
A dataframe with GTF annotation information
}
\description{
This function loads a GTF/GFF annotation into a dataframe.
}
\examples{
gtf_df <- load_gtf_annotation(gtf_annotation_path)
}
