% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{load_gtf_annotation}
\alias{load_gtf_annotation}
\title{Load GTF/GFF Annotation}
\usage{
load_gtf_annotation(path)
}
\arguments{
\item{path}{Path to GTF/GFF/GFF3 annotation}
}
\value{
A dataframe with GTF/GFF/GFF3 annotation information
}
\description{
This function loads a GTF/GFF/GFF3 annotation into a dataframe.
}
\examples{
gtf_df <- load_gtf_annotation(path/to/gtf/annotation)
}
