% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{load_kegg_annotation}
\alias{load_kegg_annotation}
\title{Load KEGG Annotation}
\usage{
load_kegg_annotation(path)
}
\arguments{
\item{path}{Path to KEGG annotation}
}
\value{
A dataframe with KEGG annotation information
}
\description{
This function loads a KEGG annotation into a dataframe.
}
\examples{
kegg_annotation_df <- load_kegg_annotation(path/to/kegg/annotation)
}
