% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfea.R
\name{transcript2protein_id}
\alias{transcript2protein_id}
\title{Convert a Vector of Transcript IDs to Protein IDs}
\usage{
transcript2protein_id(transcript_ids, transcript2protein_id_df)
}
\arguments{
\item{transcript_ids}{A vector of transcript IDs}

\item{transcript2protein_id_df}{A dataframe with gene ID, gene name, transcript ID and protein ID information (see create_transcript2protein_id_df)}
}
\value{
A vector of protein IDs
}
\description{
This function converts a vector of transcript IDs to protein IDs, using a transcript2protein_id dataframe constructed from GTF/GFF/GFF3 annotation.
}
\examples{
protein_ids <- transcript2protein_id(transcript_ids, transcript2protein_id_df)
}
